Move `sum`, `innerp`
